refactor: Move log region to core module
