+++ lynxdb-core/src/main/java/com/bailizhang/lynxdb/core/log/LogRegion.java
@@ -1,4 +1,4 @@
-package com.bailizhang.lynxdb.raft.log;
+package com.bailizhang.lynxdb.core.log;
 
 import com.bailizhang.lynxdb.core.utils.BufferUtils;
 import com.bailizhang.lynxdb.core.utils.FileUtils;
