--- lynxdb-core/src/main/java/com/bailizhang/lynxdb/core/logger/Logger.java
@@ -1,7 +1,0 @@
-package com.bailizhang.lynxdb.core.logger;
-
-public class Logger {
-    public void info(String message) {
-
-    }
-}
