+++ lynxdb-core/src/test/java/com/bailizhang/lynxdb/core/log/LogRegionTest.java
@@ -1,4 +1,4 @@
-package com.bailizhang.lynxdb.raft.log;
+package com.bailizhang.lynxdb.core.log;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -6,8 +6,6 @@ import org.junit.jupiter.api.Test;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class LogRegionTest {
     private static final String BASE_DIR = System.getProperty("user.dir") + "/logs";
