+++ lynxdb-lsm-tree/src/main/java/com/bailizhang/lynxdb/lsmtree/common/Options.java
@@ -1,10 +1,17 @@
 package com.bailizhang.lynxdb.lsmtree.common;
 
 public class Options {
+    private static final int DEFAULT_MEM_TABLE_SIZE = 2000;
+
     private boolean createColumnFamilyIfNotExisted = false;
+    private final int memTableSize;
 
     public Options() {
+        this(DEFAULT_MEM_TABLE_SIZE);
+    }
 
+    public Options(int memTableSize) {
+        this.memTableSize = memTableSize;
     }
 
     public void createColumnFamilyIfNotExisted(boolean val) {
@@ -13,5 +20,9 @@ public class Options {
 
     public boolean createColumnFamilyIfNotExisted() {
         return createColumnFamilyIfNotExisted;
+    }
+
+    public int memTableSize() {
+        return memTableSize;
     }
 }
