+++ lynxdb-raft/src/main/java/com/bailizhang/lynxdb/raft/request/AppendEntriesArgs.java
@@ -3,7 +3,7 @@ package com.bailizhang.lynxdb.raft.request;
 import com.bailizhang.lynxdb.core.common.BytesList;
 import com.bailizhang.lynxdb.core.common.BytesListConvertible;
 import com.bailizhang.lynxdb.core.utils.BufferUtils;
-import com.bailizhang.lynxdb.raft.log.LogEntry;
+import com.bailizhang.lynxdb.core.log.LogEntry;
 import com.bailizhang.lynxdb.socket.client.ServerNode;
 
 import java.nio.ByteBuffer;
